refactor(frontend): move listener types to lib
