+++ src/frontend/src/icp/services/worker.icrc-wallet.services.ts
@@ -3,7 +3,7 @@ import {
 	onTransactionsCleanUp
 } from '$icp/services/ic-transactions.services';
 import type { IcToken } from '$icp/types/ic';
-import type { WalletWorker } from '$icp/types/ic-listener';
+import type { WalletWorker } from '$lib/types/listener';
 import type {
 	PostMessage,
 	PostMessageDataResponseError,
